Big Decimals to check limits

+++ backend/ccat-revamp/user-management-service/src/main/java/com/asset/ccat/user_management/database/extractors/ProfileLimitsExtractor.java
@@ -25,7 +25,7 @@ public class ProfileLimitsExtractor implements ResultSetExtractor<HashMap<Intege
             if (profilesMap.get(limitId) == null) {
                 LkProfileLimit profileLimit = new LkProfileLimit();
                 profileLimit.setLimitId(resultSet.getInt(DatabaseStructs.ADM_PROFILE_MONETARY_LIMIT.LIMIT_ID));
-                profileLimit.setValue(resultSet.getFloat(DatabaseStructs.ADM_PROFILE_MONETARY_LIMIT.VALUE));
+                profileLimit.setValue(resultSet.getBigDecimal(DatabaseStructs.ADM_PROFILE_MONETARY_LIMIT.VALUE));
                 profileLimit.setLimitName(resultSet.getString(DatabaseStructs.LK_MONETARY_LIMITS.LIMIT_NAME));
                 if (!limitId.equals(0)) {
                     profilesMap.put(limitId, profileLimit);
